% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quotes.R
\name{add_quotes}
\alias{add_quotes}
\title{Adds quotes to a character string}
\usage{
add_quotes(x)
}
\arguments{
\item{x}{a character string (or vector)}
}
\description{
Adds quotes to a character string
}
\examples{
# Add quotes around a character string
add_quotes("A character string")

# Add quotes around a vector of strings
add_quotes(c("a", "b"))
}
