% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_language.R
\name{check_language}
\alias{check_language}
\title{Check language validity ("de", "fr", "it", "en")}
\usage{
check_language(language)
}
\arguments{
\item{language}{one of "de", "fr", "it" or "en"}
}
\description{
Check language validity ("de", "fr", "it", "en")
}
