% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_accessService.R
\name{dcat_accessService}
\alias{dcat_accessService}
\title{Create a dcat:accessService property in DCAT-AP CH standard}
\usage{
dcat_accessService(x = dcat_DataService())
}
\arguments{
\item{x}{things to include in a dcat:accessService,
typically one or more dcat:DataService}
}
\description{
Create a dcat:accessService property in DCAT-AP CH standard
}
