% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_accessURL.R
\name{dcat_accessURL}
\alias{dcat_accessURL}
\title{Create a dcat:accessURL property in DCAT-AP CH standard}
\usage{
dcat_accessURL(
  access_url = "https://www3.ti.ch/DFE/DR/USTAT/allegati/cubo/cubi_POL_01_csv.zip"
)
}
\arguments{
\item{access_url}{url of the resource

Mandatory property of \code{dcat:Distribution} (\code{\link[=dcat_distribution]{dcat_distribution()}})}
}
\description{
Create a dcat:accessURL property in DCAT-AP CH standard
}
\examples{
dcat_accessURL(access_url = "https://www3.ti.ch/DFE/DR/USTAT/allegati/cubo/cubi_POL_01_csv.zip")
}
