% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_byteSize.R
\name{dcat_byteSize}
\alias{dcat_byteSize}
\title{Create a dcat:byteSize property in DCAT-AP CH standard}
\usage{
dcat_byteSize(size = 0)
}
\arguments{
\item{size}{size of the resource (in bytes)

Property of \code{dcat:Distribution} (\code{\link[=dcat_distribution]{dcat_distribution()}})}
}
\description{
Create a dcat:byteSize property in DCAT-AP CH standard
}
