% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_catalog.R
\name{dcat_catalog}
\alias{dcat_catalog}
\title{Create a DCAT-AP CH Catalog (dcat:Catalog)}
\usage{
dcat_catalog(
  catalog_endpoint = "https://swisstopo/catalog-endpoint.rdf",
  datasets = "<dcat:dataset rdf:resource=\\"https://swisstopo/123\\"/>"
)
}
\arguments{
\item{catalog_endpoint}{URL/URI of the catalog endpoint}

\item{datasets}{character vector with one or more dcat:dataset(s), see
\code{\link[=dcat_dataset]{dcat_dataset()}}}
}
\description{
Create a DCAT-AP CH Catalog (dcat:Catalog)
}
