% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_contactPoint.R
\name{dcat_contactPoint}
\alias{dcat_contactPoint}
\title{Create a dcat:contactPoint property in DCAT-AP CH standard}
\usage{
dcat_contactPoint(
  nodeID = paste(sample(letters), collapse = ""),
  email = "dfe-ustat@ti.ch"
)
}
\arguments{
\item{nodeID}{organization id (character string)}

\item{email}{email address}
}
\description{
Create a dcat:contactPoint property in DCAT-AP CH standard
}
