% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_DataService.R
\name{dcat_DataService}
\alias{dcat_DataService}
\title{Create a DCAT-AP CH DataService (dcat:DataService)}
\usage{
dcat_DataService(about = "", endpointURL = dcat_endpointURL(), ...)
}
\arguments{
\item{about}{url of the DataService}

\item{endpointURL}{enpointURL of the dataset, see dcat_endpointURL()}

\item{...}{any other property to be included to a dcat:Dataset}
}
\value{
A character vector of length greater than one.
}
\description{
Create a DCAT-AP CH DataService (dcat:DataService)
}
\details{
A dcat:DataService class, created with dcat_DataService(),
can/must have the following properties (M stands
for mandatory; R for recommended):
\itemize{
\item dcat:endpointURL, M in V2: see \code{\link[=dcat_endpointURL]{dcat_endpointURL()}}
\item dcat:contactPoint, M in V2: see \code{\link[=dcat_contactPoint]{dcat_contactPoint()}}
\item dct:publisher, M in V2: see \code{\link[=dct_publisher]{dct_publisher()}}
\item dct:title, M in V2: see \code{\link[=dct_title]{dct_title()}}
\item dct:endpointDescription, R in V2: see \code{\link[=dcat_endpointDescription]{dcat_endpointDescription()}}
\item dct:license, R in V2: see \code{\link[=dct_license]{dct_license()}} and/or \code{\link[=dct_license2]{dct_license2()}}
\item dcat:servesDataset, R in V2: not yet implemented
}
}
