% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_enpointDescription.R
\name{dcat_endpointDescription}
\alias{dcat_endpointDescription}
\title{Create a dcat:endpointDescription property in DCAT-AP CH standard}
\usage{
dcat_endpointDescription(x = "")
}
\arguments{
\item{x}{a url}
}
\description{
Create a dcat:endpointDescription property in DCAT-AP CH standard
}
