% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_endpointURL.R
\name{dcat_endpointURL}
\alias{dcat_endpointURL}
\title{Create a dcat:endpointURL property in DCAT-AP CH standard}
\usage{
dcat_endpointURL(x = "")
}
\arguments{
\item{x}{a url}
}
\description{
Create a dcat:endpointURL property in DCAT-AP CH standard
}
