% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_keyword.R
\name{dcat_keyword}
\alias{dcat_keyword}
\title{Create a dcat:keyword property in DCAT-AP CH standard}
\usage{
dcat_keyword(x = "", language = c("de", "fr", "it", "en"))
}
\arguments{
\item{x}{keyword}

\item{language}{one of "de", "fr", "it" or "en"}
}
\description{
Create a dcat:keyword property in DCAT-AP CH standard
}
\examples{
dcat_keyword()
dcat_keyword("Cani", language = "it")
dcat_keyword("Hunde", language = "de")

}
