% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_landingPage.R
\name{dcat_landingPage}
\alias{dcat_landingPage}
\title{Create a dcat:landingPage property in DCAT-AP CH standard}
\usage{
dcat_landingPage(x = "https://www.example.com")
}
\arguments{
\item{x}{an url (character string). Default: https://www.example.com}
}
\description{
Create a dcat:landingPage property in DCAT-AP CH standard
}
\examples{
dcat_landingPage()

}
