% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_servesDataset.R
\name{dcat_servesDataset}
\alias{dcat_servesDataset}
\title{Create a dcat:servesDataset property in DCAT-AP CH standard}
\usage{
dcat_servesDataset(x = "https://example.com")
}
\arguments{
\item{x}{a url}
}
\description{
Create a dcat:servesDataset property in DCAT-AP CH standard
}
\examples{
dcat_servesDataset("https://example.com")

}
