% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_service.R
\name{dcat_service}
\alias{dcat_service}
\title{Create a dcat:service property in DCAT-AP CH standard}
\usage{
dcat_service(x = "")
}
\arguments{
\item{x}{things to include in a dcat:service, typically one or more dcat:DataService}
}
\description{
Create a dcat:service property in DCAT-AP CH standard
}
