% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_theme.R
\name{dcat_theme}
\alias{dcat_theme}
\title{Create a dcat:theme property in DCAT-AP CH standard}
\usage{
dcat_theme(x = "http://dcat-ap.ch/vocabulary/themes/work")
}
\arguments{
\item{x}{a theme url}
}
\description{
Create a dcat:theme property in DCAT-AP CH standard
}
\examples{
dcat_theme(x = "http://dcat-ap.ch/vocabulary/themes/work")

}
