% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_themeTaxonomy.R
\name{dcat_themeTaxonomy}
\alias{dcat_themeTaxonomy}
\title{Create a dcat:themeTaxonomy property in DCAT-AP CH standard}
\usage{
dcat_themeTaxonomy(
  x = "http://publications.europa.eu/resource/authority/data-theme"
)
}
\arguments{
\item{x}{an url}
}
\description{
Create a dcat:themeTaxonomy property in DCAT-AP CH standard
}
\examples{
dcat_themeTaxonomy(x = "http://publications.europa.eu/resource/authority/data-theme")

}
