% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcatapchr-package.R
\docType{package}
\name{dcatapchr-package}
\alias{dcatapchr}
\alias{dcatapchr-package}
\title{dcatapchr: Create DCAT-AP CH Metadata Files}
\description{
Create DCAT-AP CH metadata files, typically in rdf format.
}
\author{
\strong{Maintainer}: Sandro Burri \email{gibo.gaf@gmail.com}

}
\keyword{internal}
