% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_accrualPeriodicity.R
\name{dct_accrualPeriodicity}
\alias{dct_accrualPeriodicity}
\title{Create a dct:accrualPeriodicity property in DCAT-AP CH standard}
\usage{
dct_accrualPeriodicity(
  freq = c("https://publications.europa.eu/resource/authority/frequency/ANNUAL",
    "https://publications.europa.eu/resource/authority/frequency/QUARTERLY",
    "https://publications.europa.eu/resource/authority/frequency/MONTHLY",
    "https://publications.europa.eu/resource/authority/frequency/DAILY")
)
}
\arguments{
\item{freq}{frequency with which a dataset is updated (see \url{https://publications.europa.eu/resource/authority/frequency})}
}
\description{
Create a dct:accrualPeriodicity property in DCAT-AP CH standard
}
\examples{
dct_accrualPeriodicity(
  freq = "https://publications.europa.eu/resource/authority/frequency/ANNUAL"
)

}
