% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_description.R
\name{dct_description}
\alias{dct_description}
\title{Create a dct:description property in DCAT-AP CH standard}
\usage{
dct_description(description = "", language = "it")
}
\arguments{
\item{description}{character string with a description}

\item{language}{language of the title (one of "it", "de", "fr", "en")}
}
\description{
Create a dct:description property in DCAT-AP CH standard
}
\examples{
tmp <- paste0(
  "This is a description of a certain ",
  "distribution of a dataset"
)

dct_description(description = tmp, language = "en")

rm(tmp)
}
