% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_format.R
\name{dct_format}
\alias{dct_format}
\title{Create a dct:format property in DCAT-AP CH standard}
\usage{
dct_format(
  x = "http://publications.europa.eu/resource/authority/file-type/CSV"
)
}
\arguments{
\item{x}{a filetype resource url/uri. See https://publications.europa.eu/resource/authority/file-type}
}
\description{
Create a dct:format property in DCAT-AP CH standard
}
