% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_identifier.R
\name{dct_identifier}
\alias{dct_identifier}
\title{Create a dct:identifier property in DCAT-AP CH standard}
\usage{
dct_identifier(identifier = "xyz456@cantone_ticino")
}
\arguments{
\item{identifier}{character string with and identifier}
}
\description{
Create a dct:identifier property in DCAT-AP CH standard
}
\examples{
dct_identifier(identifier = "ti-ustat-cubi_pol_01@ustat")

}
