% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_issued.R
\name{dct_issued}
\alias{dct_issued}
\title{Create a dct:issued property in DCAT-AP CH standard}
\usage{
dct_issued(date = Sys.Date(), hour_minutes_seconds = "00:00:00")
}
\arguments{
\item{date}{date or character vector in yyyy-mm-dd format}

\item{hour_minutes_seconds}{hour, minutes and seconds (default: "00:00:00")}
}
\description{
Create a dct:issued property in DCAT-AP CH standard
}
