% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_license.R
\name{dct_license}
\alias{dct_license}
\title{Create a dct:license property in DCAT-AP CH standard}
\usage{
dct_license(about = "http://dcat-ap.ch/vocabulary/licenses/terms_open")
}
\arguments{
\item{about}{a dct:license uri. See https://www.dcat-ap.ch/vocabulary/licenses/20210623.html}
}
\description{
Create a dct:license property in DCAT-AP CH standard
}
