% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_modified.R
\name{dct_modified}
\alias{dct_modified}
\title{Create a dct:modified property in DCAT-AP CH standard}
\usage{
dct_modified(date = Sys.Date(), hour_minutes_seconds = "00:00:00")
}
\arguments{
\item{date}{date or character vector in yyyy-mm-dd format}

\item{hour_minutes_seconds}{hour, minutes and seconds (default: "00:00:00")}
}
\description{
Create a dct:modified property in DCAT-AP CH standard
}
\examples{
dct_modified(date = "2001-12-31", hour_minutes_seconds = "10:00:00")

}
