% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_publisher.R
\name{dct_publisher}
\alias{dct_publisher}
\title{Create a dct:publisher property in DCAT-AP CH standard}
\usage{
dct_publisher(
  publisher = "https://www.ti.ch/ustat",
  foaf_name = "Ufficio di statistica del Cantone Ticino (Ustat)"
)
}
\arguments{
\item{publisher}{character string with a publisher (usually an URL)}

\item{foaf_name}{name of the organization/publisher}
}
\description{
Create a dct:publisher property in DCAT-AP CH standard
}
