% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_rights.R
\name{dct_rights}
\alias{dct_rights}
\title{Create a dct:rights property in DCAT-AP CH standard}
\usage{
dct_rights(
  license = c("NonCommercialAllowed-CommercialAllowed-ReferenceNotRequired",
    "NonCommercialAllowed-CommercialAllowed-ReferenceRequired",
    "NonCommercialAllowed-CommercialWithPermission-ReferenceNotRequired",
    "NonCommercialAllowed-CommercialWithPermission-ReferenceRequired")
)
}
\arguments{
\item{license}{one of the licenses allowed in opendata.swiss

See also https://dcat-ap.ch/vocabulary/licenses/20210623.html#terms_open for a
list of possible licenses.}
}
\description{
Create a dct:rights property in DCAT-AP CH standard
}
\examples{
dct_rights(
  license = "NonCommercialAllowed-CommercialAllowed-ReferenceRequired"
)

}
