% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_spatial.R
\name{dct_spatial}
\alias{dct_spatial}
\title{Create a dct:spatial property in DCAT-AP CH standard}
\usage{
dct_spatial(x = "Switzerland")
}
\arguments{
\item{x}{spatial reference (default: "Switzerland")}
}
\description{
Create a dct:spatial property in DCAT-AP CH standard
}
