% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_temporal.R
\name{dct_temporal}
\alias{dct_temporal}
\title{Create a dct:temporal property in DCAT-AP CH standard}
\usage{
dct_temporal(startDate = Sys.Date() - 1, endDate = Sys.Date())
}
\arguments{
\item{startDate}{start date (default: Sys.Date() - 1). Date in yyyy-mm-dd format}

\item{endDate}{end date (default: Sys.Date()). Date in yyyy-mm-dd format}
}
\description{
Create a dct:temporal property in DCAT-AP CH standard
}
\examples{
dct_temporal(
  startDate = "2000-12-31",
  endDate = "2010-12-31"
)

}
