% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_title.R
\name{dct_title}
\alias{dct_title}
\title{Create a dct:title property in DCAT-AP CH standard}
\usage{
dct_title(title = "", language = "it")
}
\arguments{
\item{title}{character string with a title}

\item{language}{language of the title (one of "it", "de", "fr", "en")

Mandatory property of \code{dcat:Catalog} (V2) (\code{\link[=dcat_catalog]{dcat_catalog()}}) and
\code{dcat:Dataset} (\code{\link[=dcat_dataset]{dcat_dataset()}}), recommended property of
\code{dcat:Distribution} (\code{\link[=dcat_distribution]{dcat_distribution()}})}
}
\description{
Create a dct:title property in DCAT-AP CH standard
}
\examples{
dct_title(title = "This is a title", language = "en")


}
