% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_gen_tag.R
\name{f_gen_tag}
\alias{f_gen_tag}
\title{A closure to create a function that generates a tag like \if{html}{\out{<tag>}}...\if{html}{\out{</tag>}}}
\usage{
f_gen_tag(tag = "dct:identifier")
}
\arguments{
\item{tag}{character string with the desired tag}
}
\description{
A closure to create a function that generates a tag like \if{html}{\out{<tag>}}...\if{html}{\out{</tag>}}
}
\examples{
# Create an f function to generate a dct:identifier tag
f <- f_gen_tag(tag = "dct:identifier")
f(x = "an_identifier")


}
