% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_gen_tag.R
\name{f_gen_tag2}
\alias{f_gen_tag2}
\title{A closure to create a function that generates a tag like \if{html}{\out{<tag rdf:resource="..." />}}}
\usage{
f_gen_tag2(tag = "dcat:accessURL", rdf_resource = "rdf:resource")
}
\arguments{
\item{tag}{character string with the desired tag}

\item{rdf_resource}{character string that comes afer the tag (default: "rdf:resource")}
}
\description{
A closure to create a function that generates a tag like \if{html}{\out{<tag rdf:resource="..." />}}
}
\examples{
# Create an f function to generate a dcat:accessURL tag
f <- f_gen_tag2(tag = "dcat:accessURL")
f(x = "https://www.example.com")


}
