% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foaf_homepage.R
\name{foaf_homepage}
\alias{foaf_homepage}
\title{Create a foaf:homepage tag}
\usage{
foaf_homepage(x = "http://www.example.com")
}
\arguments{
\item{x}{an url of a homepage}
}
\description{
Create a foaf:homepage tag
}
