% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tag.R
\name{gen_tag}
\alias{gen_tag}
\title{Generate a tag like \if{html}{\out{<tag>}}...\if{html}{\out{</tag>}}}
\usage{
gen_tag(tag = "dct:identifier", x = "...")
}
\arguments{
\item{tag}{character string with the desired tag}

\item{x}{the content to be encapsulated inside the tag}
}
\description{
Generate a tag like \if{html}{\out{<tag>}}...\if{html}{\out{</tag>}}
}
