% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tag.R
\name{gen_tag2}
\alias{gen_tag2}
\title{Generate a tag like \if{html}{\out{<tag rdf:resource="..." />}}}
\usage{
gen_tag2(
  tag = "dcat:accessURL",
  x = "https://example.com",
  rdf_resource = "rdf:resource"
)
}
\arguments{
\item{tag}{character string with the desired tag}

\item{x}{the content to be encapsulated inside the tag. Note that this
content is enclosed in quotation marks (by the function)}

\item{rdf_resource}{character string that}
}
\description{
Generate a tag like \if{html}{\out{<tag rdf:resource="..." />}}
}
