% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent.R
\name{indent}
\alias{indent}
\title{Indent a character vector by some spaces}
\usage{
indent(x, nspaces = 2L)
}
\arguments{
\item{x}{a character vector}

\item{nspaces}{number of desired spaces to indent (default: 2)}
}
\description{
Indent a character vector by some spaces
}
