% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_dataset.R
\name{rdf_dataset}
\alias{rdf_dataset}
\title{Prepare a dataset to be exported in rdf/xml}
\usage{
rdf_dataset(
  dataset = dcat_dataset(),
  namespaces = c("xmlns:foaf=\\"http://xmlns.com/foaf/0.1/\\"",
    "xmlns:rdfs=\\"http://www.w3.org/2000/01/rdf-schema#\\"",
    "xmlns:hydra=\\"http://www.w3.org/ns/hydra/core#\\"",
    "xmlns:rdf=\\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\\"",
    "xmlns:dcat=\\"http://www.w3.org/ns/dcat#\\"",
    "xmlns:dct=\\"http://purl.org/dc/terms/\\"",
    "xmlns:vcard=\\"http://www.w3.org/2006/vcard/ns#\\"",
    "xmlns:xsd=\\"http://www.w3.org/2001/XMLSchema#\\"",
    "xmlns:schema=\\"http://schema.org/\\"")
)
}
\arguments{
\item{dataset}{a dataset created with \code{\link[=dcat_dataset]{dcat_dataset()}}}

\item{namespaces}{character vector with xml/rdf namespaces}
}
\description{
Prepare a dataset to be exported in rdf/xml
}
