% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_namespaces.R
\name{rdf_namespaces}
\alias{rdf_namespaces}
\title{Create a character vector with some xml/rdf namespaces}
\usage{
rdf_namespaces(
  namespaces = c("xmlns:foaf=\\"http://xmlns.com/foaf/0.1/\\"",
    "xmlns:rdfs=\\"http://www.w3.org/2000/01/rdf-schema#\\"",
    "xmlns:hydra=\\"http://www.w3.org/ns/hydra/core#\\"",
    "xmlns:rdf=\\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\\"",
    "xmlns:dcat=\\"http://www.w3.org/ns/dcat#\\"",
    "xmlns:dct=\\"http://purl.org/dc/terms/\\"",
    "xmlns:vcard=\\"http://www.w3.org/2006/vcard/ns#\\"",
    "xmlns:xsd=\\"http://www.w3.org/2001/XMLSchema#\\"",
    "xmlns:schema=\\"http://schema.org/\\""),
  nspaces = 2L
)
}
\arguments{
\item{namespaces}{character vector with xml/rdf namespaces}

\item{nspaces}{number of spaces to indent (default: 2)}
}
\description{
Create a character vector with some xml/rdf namespaces
}
