% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_unix.R
\name{write_unix}
\alias{write_unix}
\title{Write a character vector to a file with unix file endings and UTF-8 encoding}
\usage{
write_unix(x, filename)
}
\arguments{
\item{x}{a character vector}

\item{filename}{filename}
}
\description{
Write a character vector to a file with unix file endings and UTF-8 encoding
}
