% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_catalog.R
\name{dcat_catalog}
\alias{dcat_catalog}
\title{Create a DCAT-AP CH Catalog (dcat:Catalog)}
\usage{
dcat_catalog(
  catalog_endpoint = "https://swisstopo/catalog-endpoint.rdf",
  datasets = "<dcat:dataset rdf:resource=\\"https://swisstopo/123\\"/>",
  ...
)
}
\arguments{
\item{catalog_endpoint}{URL/URI of the catalog endpoint}

\item{datasets}{character vector with one or more dcat:dataset(s), see
\code{\link[=dcat_dataset]{dcat_dataset()}}}

\item{...}{any other property to be included to a dcat:Catalog}
}
\description{
Create a DCAT-AP CH Catalog (dcat:Catalog)
}
\examples{
dcat_catalog()
dcat_catalog(catalog_endpoint = "")


# Add some properties (...)
dcat_catalog(
  title = dct_title("Catalogo Ustat", language = "it"),
  description = dct_description("Descrizione catalogo", language = "it")
)

}
