% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_catalog.R
\name{dcat_catalog}
\alias{dcat_catalog}
\title{Create a DCAT-AP CH Catalog (dcat:Catalog)}
\usage{
dcat_catalog(
  catalog_endpoint = "https://swisstopo/catalog-endpoint.rdf",
  datasets = "<dcat:dataset rdf:resource=\\"https://swisstopo/123\\"/>",
  ...
)
}
\arguments{
\item{catalog_endpoint}{URL/URI of the catalog endpoint}

\item{datasets}{character vector with one or more dcat:dataset(s), see
\code{\link[=dcat_dataset]{dcat_dataset()}}}

\item{...}{any other property to be included to a dcat:Catalog}
}
\value{
A character vector of length greater than one,
where each element is a line of the catalog in rdf
format. The catalog can be exported in a file with
unix file endings and UTF-8 encoding with
\code{\link[=write_unix]{write_unix()}}.
}
\description{
Create a DCAT-AP CH Catalog (dcat:Catalog)
}
\details{
A dcat:Catalog class, created with dcat_catalog(),
can/must have the following properties (M stands
for mandatory; R for recommended):
\itemize{
\item dcat:dataset, M in V1, R in V2: see \code{\link[=dcat_dataset]{dcat_dataset()}}
\item dct:description, M in V2: see \code{\link[=dct_description]{dct_description()}}
\item dct:publisher, M in V2: see \code{\link[=dct_publisher]{dct_publisher()}}
\item dct:title, M in V2: see \code{\link[=dct_title]{dct_title()}}
\item foaf:homepage, R in V2: see \code{\link[=foaf_homepage]{foaf_homepage()}}
\item dct:language, R in V2: see \code{\link[=dct_language]{dct_language()}}
\item dct:license, R in V2: see \code{\link[=dct_license]{dct_license()}}
\item dct:issued, R in V2: see \code{\link[=dct_issued]{dct_issued()}}
\item dct:rights, R in V2: see \code{\link[=dct_rights]{dct_rights()}}
\item dct:spatial, R in V2: see \code{\link[=dct_spatial]{dct_spatial()}}
\item dcat:themeTaxonomy, R in V2: see \code{\link[=dcat_themeTaxonomy]{dcat_themeTaxonomy()}}
\item dct:modified, R in V2: see \code{\link[=dct_modified]{dct_modified()}}
\item dcat:service, R in V2: see \code{\link[=dcat_service]{dcat_service()}}
}
}
\examples{
dcat_catalog()
dcat_catalog(catalog_endpoint = "")


# Add some properties (...)
dcat_catalog(
  title = dct_title("Catalogo Ustat", language = "it"),
  description = dct_description("Descrizione catalogo", language = "it")
)

}
