% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_contactPoint.R
\name{dcat_contactPoint}
\alias{dcat_contactPoint}
\title{Create a dcat:contactPoint property in DCAT-AP CH standard}
\usage{
dcat_contactPoint(
  nodeID = paste(sample(letters), collapse = ""),
  email = "dfe-ustat@ti.ch",
  fn = "Ufficio di statistica del Canton Ticino (Ustat)"
)
}
\arguments{
\item{nodeID}{organization id (character string)}

\item{email}{email address}

\item{fn}{name of the organization

Mandatory property of \code{dcat:Dataset} (\code{\link[=dcat_dataset]{dcat_dataset()}})}
}
\description{
Create a dcat:contactPoint property in DCAT-AP CH standard
}
\examples{
# Create a dcat:contactPoint tag
dcat_contactPoint(
  nodeID = "organization_id",
  email = "dfe-ustat@ti.ch",
  fn = "Ufficio di statistica del Canton Ticino (Ustat)"
)
}
