% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_dataset.R
\name{dcat_dataset}
\alias{dcat_dataset}
\title{Create a DCAT-AP CH Catalog (dcat:dataset)}
\usage{
dcat_dataset(
  title = dct_title(),
  description = dct_description(),
  publisher = dct_publisher(),
  contactPoint = dcat_contactPoint(),
  identifier = dct_identifier(),
  distribution = dcat_distribution(),
  ...
)
}
\arguments{
\item{title}{title of the dataset, see \code{\link[=dct_title]{dct_title()}}. Can be multilingual.}

\item{description}{description of the dataset, see \code{\link[=dct_description]{dct_description()}}. Can be multilingual.}

\item{publisher}{publisher of the dataset, see \code{\link[=dct_publisher]{dct_publisher()}}.}

\item{contactPoint}{contactPoint of the dataset, see \code{\link[=dcat_contactPoint]{dcat_contactPoint()}}.}

\item{identifier}{identifier of the dataset, see \code{\link[=dct_identifier]{dct_identifier()}}.}

\item{distribution}{one or more distributions, see \code{\link[=dcat_distribution]{dcat_distribution()}}}

\item{...}{additional arguments (not yet implemented)}
}
\description{
A dcat:dataset must have one or more dcat:distribution(s).
}
