% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_distribution.R
\name{dcat_distribution}
\alias{dcat_distribution}
\title{Create a dcat:distribution/dcat:Distribution property}
\usage{
dcat_distribution(
  issued = dct_issued(),
  access_url = dcat_accessURL(),
  license = dct_license2(),
  ...
)
}
\arguments{
\item{issued}{dct:issued property, see \code{\link[=dct_issued]{dct_issued()}}.}

\item{access_url}{dcat:accessURL property, see \code{\link[=dcat_accessURL]{dcat_accessURL()}}.}

\item{license}{dct:license property, see \code{\link[=dct_license]{dct_license()}}.}

\item{...}{other arguments (not yet implemented)

Mandatory in V1, recommended in V2 (for \code{dcat:Dataset}, \code{\link[=dcat_dataset]{dcat_dataset()}})}
}
\description{
Create a dcat:distribution/dcat:Distribution property
}
