% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_distribution.R
\name{dcat_distribution}
\alias{dcat_distribution}
\title{Create a dcat:distribution/dcat:Distribution property}
\usage{
dcat_distribution(
  issued = dct_issued(),
  access_url = dcat_accessURL(),
  rights = dct_rights(),
  ...
)
}
\arguments{
\item{issued}{dct:issued property, see \code{\link[=dct_issued]{dct_issued()}}.}

\item{access_url}{dcat:accessURL property, see \code{\link[=dcat_accessURL]{dcat_accessURL()}}.}

\item{rights}{dct:rights property, see \code{\link[=dct_rights]{dct_rights()}}.}

\item{...}{other arguments (not yet implemented)}
}
\description{
Create a dcat:distribution/dcat:Distribution property
}
