% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcat_downloadURL.R
\name{dcat_downloadURL}
\alias{dcat_downloadURL}
\title{Create a dcat:downloadURL property in DCAT-AP CH standard}
\usage{
dcat_downloadURL(
  download_url = "https://www3.ti.ch/DFE/DR/USTAT/allegati/cubo/cubi_POL_01_csv.zip"
)
}
\arguments{
\item{download_url}{url of the resource

Property of \code{dcat:Distribution} (\code{\link[=dcat_distribution]{dcat_distribution()}})}
}
\description{
Create a dcat:downloadURL property in DCAT-AP CH standard
}
