% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_language.R
\name{dct_language}
\alias{dct_language}
\title{Create a dct:language property in DCAT-AP CH standard}
\usage{
dct_language(language = c("de", "fr", "it", "en"))
}
\arguments{
\item{language}{two-letters abbreviation of the language (one of "de", "fr", "it" or "en")}
}
\description{
Create a dct:language property in DCAT-AP CH standard
}
\examples{
dct_language(language = "de")
dct_language(language = "fr")
dct_language(language = "it")
dct_language(language = "en")

}
