% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_title.R
\name{dct_title}
\alias{dct_title}
\title{Create a dct:title property in DCAT-AP CH standard}
\usage{
dct_title(title = "", language = "it")
}
\arguments{
\item{title}{character string with a title}

\item{language}{language of the title (one of "it", "de", "fr", "en")}
}
\description{
Create a dct:title property in DCAT-AP CH standard
}
